+++ main.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   main.c                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mrosario <mrosario@student.42.fr>          +#+  +:+       +#+        */
+/*   By: miki <miki@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/24 18:17:50 by mrosario          #+#    #+#             */
-/*   Updated: 2021/03/02 18:14:05 by mrosario         ###   ########.fr       */
+/*   Updated: 2021/03/08 23:02:14 by miki             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,19 +23,19 @@
 ** NOT BEING USED...
 */
 
-char	*find(char *line, char *line_end, char c)
-{
-	if (line)
-	{
-		if (line_end && line_end > line)
-			while (line < line_end && *line != c)
-				line++;
-		else
-			while (*line && *line != c)
-				line++;
-	}
-	return (line);
-}
+// char	*find(char *line, char *line_end, char c)
+// {
+// 	if (line)
+// 	{
+// 		if (line_end && line_end > line)
+// 			while (line < line_end && *line != c)
+// 				line++;
+// 		else
+// 			while (*line && *line != c)
+// 				line++;
+// 	}
+// 	return (line);
+// }
 
 /*
 ** This function defines a buffer for reading from STDIN in steps determined by
@@ -55,42 +55,44 @@ char	*find(char *line, char *line_end, char c)
 ** newline or EOF it will hang again pending another newline or EOF. Newline
 ** counts as a character.
 **
-** Note that read will NOT clear the high bytes of the memory location passed
-** as the second argument if the chunk to be copied is smaller than specified
-** in the third argument, so the value copied will not be as expected unless
-** the high bytes are zeroed.
+** This function has been rewritten from the initial loop that more closely
+** mirrored an online tutorial to better use the read buffer. The read function
+** will now read up to READLINE_BUFSIZE bytes from stdout at a time and copy
+** them to the buffer. The number of bytes read are returned and stored in the
+** size variable.
+**
+** If READLINE_BUFSIZE bytes have been copied to the buffer and EOF or new line
+** is not found at buffer[size - 1] (transforming size to position to point to
+** the last copied byte in the buffer), there is more left to be copied, so we
+** reallocate the buffer, increasing its size by a further READLINE_BUFSIZE
+** bytes, and then read again, adding the number of bytes read to the prior
+** number of bytes read. Once EOF or new line is found, it is replaced with a
+** null terminator and the function returns the address of the buffer for
+** parsing.
+**
 */
 
 char	*micli_readline(t_micli *micli)
 {
 	size_t	size;
+	size_t	bufsize;
 
 	size = 0;
-	micli->bufsize = READLINE_BUFSIZE;
-	micli->position = 0;
-	micli->buffer = clean_calloc(micli->bufsize, sizeof(char), micli);
+	bufsize = READLINE_BUFSIZE;
+	micli->buffer = clean_calloc(bufsize, sizeof(char), micli);
 	while (1)
 	{
-		size += read(STDIN_FILENO, &micli->c, 1); //Si no ha leído nada se comporta como salto de línea?? O_O Lee último carácter, salto de línea? Se queda colgado en read???
-		//If we read EOF or newline was input by user, null terminate buffer, print and return it.
-		if (micli->c == EOF || micli->c == '\n')
+		size += read(STDIN_FILENO, &micli->buffer[size], READLINE_BUFSIZE);
+		//If we read EOF or newline was input by user, null terminate buffer.
+		if (micli->buffer[size - 1] == EOF || micli->buffer[size - 1] == '\n')
 		{
-			micli->buffer[micli->position] = '\0';
-			//ft_printf("%s\n", micli->buffer); //esto imprime último comando a stdout
+			micli->buffer[size - 1] = '\0'; //Convert EOF or '\n' to null termination. We'll need to handle EOF differently as it means ctrl'/'
+			//ft_printf("%u, %s\n", size, micli->buffer); //esto imprime última línea a stdout
 			return (micli->buffer);
 		}
-		//Otherwise, add the character to our buffer and advance one byte.
-		else
-			micli->buffer[micli->position] = micli->c;
-		micli->position++;
-		//If after advancing we exceed the allocated buffer size, expand buffer by READLINE_BUFSIZE bytes and reallocate.
-		//Podemos usar funciones de libft para esto, para de momento lo hago un poco chapuceramente. ;p
-		if (micli->position >= micli->bufsize)
-		{
-			micli->bufsize += READLINE_BUFSIZE;
-			if (!(micli->buffer = ft_realloc(micli->buffer, micli->bufsize, micli)))
-				exit_failure(micli);
-		}
+		bufsize += READLINE_BUFSIZE;
+		if (!(micli->buffer = ft_realloc(micli->buffer, bufsize, micli)))
+			exit_failure(micli);		
 	}
 }
 
@@ -102,7 +104,7 @@ char	micli_loop(t_micli *micli)
 
 	while (!shutdown)//no parece que esté usando shutdown...
 	{
-		write(1, "🚀 ", 6);
+		write(STDOUT_FILENO, "🚀 ", 5);
 		micli->buffer = micli_readline(micli);//this is redundant, as the function returns micli->buffer, leaving it here for clarity
 		process_raw_line(micli->buffer, micli);
 		micli->buffer = ft_del(micli->buffer);